Add sveltekit-typescript-showcase (#50)
Wasn't sure which category to place this. It's not really a `tutorial`, more like `unofficial documentation`, but that category does not yet exist. Feel free to move edit as you see fit.

Anyways this package is basically a high-quality documentation/reference of all the concepts that relate to `typescript` + `svelte` / `sveltekit`. Should be useful for any dev using `ts` in their `svelte` projects.